refactor: align database types with updated validation package (float64/int64)

+++ internal/database/database-general.go
@@ -505,9 +505,9 @@ func ifNullFloat64(sqlFloat sql.NullFloat64) float64 {
 }
 
 // Helper function to check for null strings and return empty if null
-func ifNullInt(sqlInt sql.NullInt64) int {
+func ifNullInt(sqlInt sql.NullInt64) int64 {
 	if sqlInt.Valid {
-		return int(sqlInt.Int64)
+		return sqlInt.Int64
 	}
 	return 0
 }
